requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0

# Data Processing
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
joblib==1.3.2

# Binance API
python-binance==1.0.19

# Database
supabase==2.3.4

# Utilities
pytz==2024.1
schedule==1.2.0

